IA-Deepseek/requirements.txt
python-dotenv
openai
websockets
sqlalchemy
pymysql
pdfplumber
fastapi
email-validator
uvicorn
passlib
bcrypt
PyJWT
python-jose
python-multipart